Initial commit
Developer should implement a action class based on abstract class
PageAction. The instance of this class is used to instanciate a
menupage class which is passed on to the MenuHandler. This is the first
menupage visible to the user and loaded by the handler

The startpage action should return an instances of a MenuPage
after the action is completed. This page will then be shown
to the user. (not implemented yet)